Updated the 'show' route for each person to display their profile, including a profile picture for each.
Added the ability to add a new person to the database, who will then be available to assign tasks to in the 'edit' tasks view.

Can view each person's individual tasks (both complete and incomplete).

Completed Final Wave of project.

Next step would be to figure out how to upload an image with the creation of the new person's profile.